requirements.txt
MouseInfo==0.1.3
mss==10.0.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python3-xlib==0.15
pytweening==1.2.0
openai==1.13.3
python-dotenv==1.0.1
